:truck: move service and client code into new architecture
