+++ mcr-core/mcr_meeting/app/db/apply_dto.py
@@ -3,17 +3,17 @@ from pydantic import BaseModel as PydanticDTO
 from mcr_meeting.app.db.db import Base as BaseORM
 
 
-def update_model(ORMmodel: BaseORM, updateDTO: PydanticDTO) -> None:
+def apply_dto(ORMmodel: BaseORM, updateDTO: PydanticDTO) -> None:
     """
-    Update an ORM model with the values from a UserUpdate model.
+    Apply every field of a DTO onto an ORM model.
     """
     for key, value in updateDTO:
         setattr(ORMmodel, key, value)
 
 
-def patch_model(ORMmodel: BaseORM, updateDTO: PydanticDTO) -> None:
+def apply_dto_patch(ORMmodel: BaseORM, updateDTO: PydanticDTO) -> None:
     """
-    Update an ORM model with the values from a UserUpdate model.
+    Apply a DTO onto an ORM model, skipping fields left unset (None).
     """
     for key, value in updateDTO:
         if value is None:
