+++ mcr-core/mcr_meeting/app/infrastructure/compute_devices.py
